refactor: build a proper Python package with poetry
