+++ .github/workflows/typecheck.yml
@@ -12,11 +12,19 @@ jobs:
     steps:
       - uses: actions/checkout@v3 # Check out the repository first.
 
-      - name: Install Python dependencies
+      - name: Set up Python 3.10.6 🐍
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.10.6"
+
+      - name: Install Poetry
+        uses: snok/install-poetry@v1
+        with:
+          virtualenvs-create: false
+
+      - name: Install Python dependencies ⬇️
         working-directory: backend
-        run: |
-          python -m pip install --upgrade pip
-          [ -f requirements.txt ] && pip install -r requirements.txt
+        run: poetry install --no-interaction
 
       - name: Install TypeScript dependencies
         working-directory: frontend
