+++ backend/decky_loader/localplatform/localplatformwin.py
