+++ backend/main.py
@@ -1,4 +1,4 @@
-# This file is needed to make the relative imports in src/ work properly.
+# This file is needed to make the relative imports in decky_loader/ work properly.
 if __name__ == "__main__":
-    from src.main import main
+    from decky_loader.main import main
     main()
