+++ backend/pyinstaller.spec
@@ -1,0 +1,31 @@
+import os
+from PyInstaller.building.build_main import Analysis
+from PyInstaller.building.api import EXE, PYZ
+from PyInstaller.utils.hooks import copy_metadata
+
+a = Analysis(
+    ['main.py'],
+    datas=[
+        ('locales', 'locales'),
+        ('../plugin', 'plugin'),
+        ('static', 'static'),
+    ] + copy_metadata('decky_loader'),
+    hiddenimports=['logging.handlers', 'sqlite3'],
+)
+pyz = PYZ(a.pure, a.zipped_data)
+
+noconsole = bool(os.getenv('DECKY_NOCONSOLE'))
+name = "PluginLoader"
+if noconsole:
+    name += "_noconsole"
+
+exe = EXE(
+    pyz,
+    a.scripts,
+    a.binaries,
+    a.zipfiles,
+    a.datas,
+    name=name,
+    upx=True,
+    console=not noconsole,
+)
