+++ backend/pyproject.toml
@@ -1,0 +1,32 @@
+[tool.poetry]
+name = "decky-loader"
+version = "0.0.0"  # the real version will be autogenerated
+description = "A plugin loader for the Steam Deck"
+license = "GPLv2"
+readme = "../README.md"
+authors = []
+packages = [{include = "decky_loader"}]
+include = ["decky_loader/static/*"]
+
+[tool.poetry.dependencies]
+python = ">=3.10,<3.13"
+
+aiohttp = "^3.8.5"
+aiohttp-jinja2 = "^1.5.1"
+aiohttp-cors = "^0.7.0"
+watchdog = "^2.1.7"
+certifi = "*"
+
+[tool.poetry.group.dev.dependencies]
+pyinstaller = "^5.13.0"
+pyright = "^1.1.335"
+
+[tool.poetry.scripts]
+decky-loader = 'decky_loader.main:main'
+
+[tool.pyright]
+strict = ["*"]
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
