--- backend/pyrightconfig.json
@@ -1,3 +1,0 @@
-{
-    "strict": ["*"]
-}
