--- backend/requirements.txt
@@ -1,5 +1,0 @@
-aiohttp==3.8.5
-aiohttp-jinja2==1.5.1
-aiohttp_cors==0.7.0
-watchdog==2.1.7
-certifi==2023.7.22
